Invisible Stairs + Extra Floors + Interact Key
Replaced the jump functionality with an interaction mechanic. While overlapping with a specific collider the player can hit spacebar to perform an interaction. Currently the only example is a set of invisible stairs that the player can traverse up and down. I also added two extra floors to showcase and use the stairs.